+++ src/routes/docs/builders/toggle/preview.svelte
@@ -11,8 +11,7 @@
 		{...$toggle}
 		aria-label="Toggle italic"
 		class="grid h-9 w-9 place-items-center items-center justify-center rounded
-	bg-white text-base leading-4 text-magnum-800 shadow-lg outline-none hover:bg-magnum-100 focus:ring
-	focus:ring-magnum-400 data-[disabled]:cursor-not-allowed data-[state=on]:bg-magnum-200
+	bg-white text-base leading-4 text-magnum-800 shadow-lg hover:bg-magnum-100 data-[disabled]:cursor-not-allowed data-[state=on]:bg-magnum-200
 	data-[state=on]:text-magnum-900"
 	>
 		<Italic />